refactor: rename fishy to holo-web-conductor throughout UI
- Switch from @zippy/fishy-client to @holo-host/web-conductor-client
- Rename ui/src/fishy/ directory to ui/src/hwc/
- FishyAppClient -> WebConductorAppClient
- waitForFishy -> waitForHolochain, isFishy -> isWebConductor
- IS_FISHY -> IS_HWC, gatewayUrl -> linkerUrl
- __GATEWAY_URL__ -> __LINKER_URL__ build constant
