+++ ui/src/utils/client.ts
@@ -1,38 +1,38 @@
 import { AdminWebsocket, CellType, AppWebsocket } from "@holochain/client";
 import WebSdkApi, { AgentState } from "@holo-host/web-sdk";
-import { FishyAppClient, waitForFishy } from "@/fishy";
+import { WebConductorAppClient, waitForHolochain } from "@/hwc";
 
 export const HOLOCHAIN_APP_ID = "mewsfeed";
 export const IS_LAUNCHER = (window as any).__HC_LAUNCHER_ENV__ !== undefined;
 export const IS_HOLO_HOSTED = import.meta.env.VITE_IS_HOLO_HOSTED;
 
-// Fishy extension detection
-declare const __GATEWAY_URL__: string;
-const GATEWAY_URL = typeof __GATEWAY_URL__ !== "undefined" ? __GATEWAY_URL__ : "http://localhost:8000";
-export let IS_FISHY = false;
+// Web Conductor extension detection
+declare const __LINKER_URL__: string;
+const LINKER_URL = typeof __LINKER_URL__ !== "undefined" ? __LINKER_URL__ : "http://localhost:8000";
+export let IS_HWC = false;
 
 export const setupHolochain = async () => {
   try {
-    // Check for fishy extension first
-    if ((window as any).holochain?.isFishy) {
-      IS_FISHY = true;
+    // Check for web conductor extension first
+    if ((window as any).holochain?.isWebConductor) {
+      IS_HWC = true;
     } else {
       try {
-        await waitForFishy(3000);
-        IS_FISHY = true;
+        await waitForHolochain(3000);
+        IS_HWC = true;
       } catch {
-        // Not fishy - fall through to normal setup
+        // Not web conductor - fall through to normal setup
       }
     }
 
-    if (IS_FISHY) {
-      console.log("Fishy extension detected, using FishyAppClient");
-      const fishyClient = await FishyAppClient.connect({
-        gatewayUrl: GATEWAY_URL,
+    if (IS_HWC) {
+      console.log("Holochain extension detected, using WebConductorAppClient");
+      const hwcClient = await WebConductorAppClient.connect({
+        linkerUrl: LINKER_URL,
         roleName: "mewsfeed",
       });
       // Cast to AppWebsocket to satisfy type checker (different @holochain/client versions)
-      return fishyClient as unknown as AppWebsocket;
+      return hwcClient as unknown as AppWebsocket;
     }
 
     let client;
