+++ ui/vite.config.ts
@@ -60,7 +60,7 @@ export default defineConfig({
     Icons({ autoInstall: true }) as PluginOption,
   ],
   define: {
-    __GATEWAY_URL__: JSON.stringify(process.env.GATEWAY_URL || "http://localhost:8000"),
+    __LINKER_URL__: JSON.stringify(process.env.LINKER_URL || "http://localhost:8000"),
   },
   build: {
     target: "es2020",
